+++ src/Shapes/rect.cpp
@@ -64,7 +64,7 @@ void Rect::draw_stroke(Frame & frame) const
 {
    if (!m_doStroke) return;
 
-   float angle = (get_point(1) - pos()).angle();
+   float angle = float((get_point(1) - pos()).angle());
    float modAngle = std::fmod(angle, twoPi);
 
    // if the rectangle is axis aligned
@@ -80,7 +80,11 @@ void Rect::draw_stroke(Frame & frame) const
       else
          tl = pos();
          
-      compute_AA_rect_stroke(frame, tl.x, tl.y, width(), height(), stroke(), strokeweight());
+      compute_AA_rect_stroke(frame, 
+         static_cast<int>(tl.x), 
+         static_cast<int>(tl.y), 
+         static_cast<int>(width()), 
+         static_cast<int>(height()), stroke(), strokeweight());
    } else {
       RectCorners crs = get_corners();
